+++ plugins/destination/kafka/client/spec.go
@@ -1,28 +1,28 @@
 package client
 
-import "fmt"
+import (
+	"fmt"
 
-type FormatType string
-
-const (
-	FormatTypeCSV  = "csv"
-	FormatTypeJSON = "json"
+	"github.com/cloudquery/filetypes"
 )
 
 type Spec struct {
-	Brokers      []string   `json:"brokers,omitempty"`
-	Format       FormatType `json:"format,omitempty"`
-	Verbose      bool       `json:"verbose,omitempty"`
-	SaslUsername string     `json:"sasl_username,omitempty"`
-	SaslPassword string     `json:"sasl_password,omitempty"`
+	Brokers      []string `json:"brokers,omitempty"`
+	Verbose      bool     `json:"verbose,omitempty"`
+	SaslUsername string   `json:"sasl_username,omitempty"`
+	SaslPassword string   `json:"sasl_password,omitempty"`
 	// This is currently only used for testing to wait for
 	// kafka cluster to be ready in GitHub actions.
 	MaxMetadataRetries int `json:"max_metadata_retries,omitempty"`
+	*filetypes.FileSpec
 }
 
-func (*Spec) SetDefaults() {
+func (s *Spec) SetDefaults() {
+	if s.FileSpec == nil {
+		s.FileSpec = &filetypes.FileSpec{}
+	}
+	s.FileSpec.SetDefaults()
 }
-
 func (s *Spec) Validate() error {
 	if len(s.Brokers) == 0 {
 		return fmt.Errorf("at least one broker is required")