+++ plugins/source/azure/resources/services/appservice/recommendations.go
@@ -15,22 +15,8 @@ func Recommendations() *schema.Table {
 		Resolver:    fetchRecommendations,
 		Description: "https://learn.microsoft.com/en-us/rest/api/appservice/recommendations/list#recommendation",
 		Multiplex:   client.SubscriptionMultiplexRegisteredNamespace("azure_appservice_recommendations", client.Namespacemicrosoft_web),
-		Transform:   transformers.TransformWithStruct(&armappservice.Recommendation{}),
-		Columns: []schema.Column{
-			{
-				Name:     "subscription_id",
-				Type:     schema.TypeString,
-				Resolver: client.ResolveAzureSubscription,
-			},
-			{
-				Name:     "id",
-				Type:     schema.TypeString,
-				Resolver: schema.PathResolver("ID"),
-				CreationOptions: schema.ColumnCreationOptions{
-					PrimaryKey: true,
-				},
-			},
-		},
+		Transform:   transformers.TransformWithStruct(&armappservice.Recommendation{}, transformers.WithPrimaryKeys("ID")),
+		Columns:     schema.ColumnList{client.SubscriptionID},
 	}
 }