+++ plugins/source/azure/resources/services/mariadb/server_configurations.go
@@ -11,16 +11,7 @@ func server_configurations() *schema.Table {
 		Name:        "azure_mariadb_server_configurations",
 		Resolver:    fetchServerConfigurations,
 		Description: "https://learn.microsoft.com/en-us/rest/api/mariadb/configurations/list-by-server?tabs=HTTP#configuration",
-		Transform:   transformers.TransformWithStruct(&armmariadb.Configuration{}),
-		Columns: []schema.Column{
-			{
-				Name:     "id",
-				Type:     schema.TypeString,
-				Resolver: schema.PathResolver("ID"),
-				CreationOptions: schema.ColumnCreationOptions{
-					PrimaryKey: true,
-				},
-			},
-		},
+		Transform:   transformers.TransformWithStruct(&armmariadb.Configuration{}, transformers.WithPrimaryKeys("ID")),
+		Columns:     schema.ColumnList{},
 	}
 }