+++ plugins/source/azure/resources/services/sql/server_blob_auditing_policies.go
@@ -12,36 +12,7 @@ func server_blob_auditing_policies() *schema.Table {
 		Name:        "azure_sql_server_blob_auditing_policies",
 		Resolver:    fetchServerBlobAuditingPolicies,
 		Description: "https://learn.microsoft.com/en-us/rest/api/sql/2021-11-01/server-blob-auditing-policies/list-by-server?tabs=HTTP#serverblobauditingpolicy",
-		Transform:   transformers.TransformWithStruct(&armsql.ServerBlobAuditingPolicy{}),
-		Columns: []schema.Column{
-			{
-				Name:     "subscription_id",
-				Type:     schema.TypeString,
-				Resolver: client.ResolveAzureSubscription,
-			},
-			{
-				Name:     "properties",
-				Type:     schema.TypeJSON,
-				Resolver: schema.PathResolver("Properties"),
-			},
-			{
-				Name:     "id",
-				Type:     schema.TypeString,
-				Resolver: schema.PathResolver("ID"),
-				CreationOptions: schema.ColumnCreationOptions{
-					PrimaryKey: true,
-				},
-			},
-			{
-				Name:     "name",
-				Type:     schema.TypeString,
-				Resolver: schema.PathResolver("Name"),
-			},
-			{
-				Name:     "type",
-				Type:     schema.TypeString,
-				Resolver: schema.PathResolver("Type"),
-			},
-		},
+		Transform:   transformers.TransformWithStruct(&armsql.ServerBlobAuditingPolicy{}, transformers.WithPrimaryKeys("ID")),
+		Columns:     schema.ColumnList{client.SubscriptionID},
 	}
 }